+++ Teichsteuerung/esp8266/teichsteuerung/teranis_code.h
@@ -89,7 +89,7 @@ void Pumpensteuerung (
   (*inst).Pumpensteuerung_P10 = I(int16_t,8) > R(int16_t,8);
   (*inst).Pumpensteuerung_P11 = (I(int16_t,8) + 2) > I(int16_t,14);
   (*inst).Pumpensteuerung_P12 = I(int16_t,8) < R(int16_t,10);
-  if (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 0)
+  if ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (0)))
   {
     if ((!FX(0,3) && (*inst).Pumpensteuerung_P0 && IX(0,0)))
     {
@@ -108,7 +108,7 @@ void Pumpensteuerung (
       (*inst).Pumpensteuerung_Op0 = ((uint8_t) (3));
     }
   }
-  else if (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 1)
+  else if ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (1)))
   {
     if (FX(0,3))
     {
@@ -121,19 +121,19 @@ void Pumpensteuerung (
       (*inst).Pumpensteuerung_Op0 = ((uint8_t) (5));
     }
   }
-  else if (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 2)
+  else if ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (2)))
   {
     (*inst).Pumpensteuerung_Op0 = ((uint8_t) (0));
   }
-  else if (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 3)
+  else if ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (3)))
   {
     (*inst).Pumpensteuerung_Op0 = ((uint8_t) (1));
   }
-  else if (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 4)
+  else if ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (4)))
   {
     (*inst).Pumpensteuerung_Op0 = ((uint8_t) (0));
   }
-  else if (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 5)
+  else if ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (5)))
   {
     if ((*inst).Pumpensteuerung_P2)
     {
@@ -152,25 +152,25 @@ void Pumpensteuerung (
       (*inst).Pumpensteuerung_Op0 = ((uint8_t) (9));
     }
   }
-  else if (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 6)
+  else if ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (6)))
   {
     (*inst).Pumpensteuerung_Op0 = ((uint8_t) (0));
   }
-  else if (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 7)
+  else if ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (7)))
   {
     if (!IX(0,0))
     {
       (*inst).Pumpensteuerung_Op0 = ((uint8_t) (4));
     }
   }
-  else if (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 8)
+  else if ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (8)))
   {
     if (!IX(0,1))
     {
       (*inst).Pumpensteuerung_Op0 = ((uint8_t) (4));
     }
   }
-  else if (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 9)
+  else if ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (9)))
   {
     (*inst).Pumpensteuerung_Op0 = ((uint8_t) (0));
   }
@@ -208,35 +208,35 @@ void Pumpensteuerung (
   /* Initialisierung wenn keine Operation gesetzt ist */
   /* Zuweisung der Operationsvariablen */
   /* Pumpe einschalten */
-  QX(0,0) = (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 1);
+  QX(0,0) = ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (1)));
   /* Störung Sensor Pumpenschacht */
-  FX(0,1) = (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 7);
+  FX(0,1) = ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (7)));
   /* Störung Sensor Filtergraben */
-  FX(0,2) = (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 8);
+  FX(0,2) = ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (8)));
   /* Zeitdifferenz berechnen */
-  _OV10 = (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 0) || (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 1);
+  _OV10 = ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (0))) || ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (1)));
   /* Zeitverhältnis mitteln */
-  _OV6 = (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 2) || (((int64_t) ((*inst).Pumpensteuerung_Op0))
-   == 6) || (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 9);
+  _OV6 = ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (2))) || ((*inst).Pumpensteuerung_Op0 == ((uint8_t)
+   (6))) || ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (9)));
   /* Zeit merken */
-  _OV7 = (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 3) || (((int64_t) ((*inst).Pumpensteuerung_Op0)) == 4) || _OV6;
-  if ((((int64_t) ((*inst).Pumpensteuerung_Op0)) == 5))
+  _OV7 = ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (3))) || ((*inst).Pumpensteuerung_Op0 == ((uint8_t) (4))) || _OV6;
+  if (((*inst).Pumpensteuerung_Op0 == ((uint8_t) (5))))
   {
     F(int16_t,10) = F(int16_t,4) - (*inst).time_last;
   }
-  if ((((int64_t) ((*inst).Pumpensteuerung_Op0)) == 3))
+  if (((*inst).Pumpensteuerung_Op0 == ((uint8_t) (3))))
   {
     F(int16_t,12) = F(int16_t,4) - (*inst).time_last;
   }
-  if ((((int64_t) ((*inst).Pumpensteuerung_Op0)) == 2))
+  if (((*inst).Pumpensteuerung_Op0 == ((uint8_t) (2))))
   {
     F(int16_t,14) = ((int16_t) ((F(int16_t,10) * 100) / (F(int16_t,10) + F(int16_t,12))));
   }
-  if ((((int64_t) ((*inst).Pumpensteuerung_Op0)) == 6))
+  if (((*inst).Pumpensteuerung_Op0 == ((uint8_t) (6))))
   {
     F(int16_t,14) = ((int16_t) ((F(int16_t,10) * 10) / (F(int16_t,10) / 10 + F(int16_t,12) / 10)));
   }
-  if ((((int64_t) ((*inst).Pumpensteuerung_Op0)) == 9))
+  if (((*inst).Pumpensteuerung_Op0 == ((uint8_t) (9))))
   {
     F(int16_t,14) = ((int16_t) (F(int16_t,10) / (F(int16_t,10) / 100 + F(int16_t,12) / 100)));
   }
@@ -291,14 +291,14 @@ void Solaranlage (
   /* Berechnung der nicht binären Prozeßvariablen */
   (*inst).Solaranlage_P0 = I(int16_t,12) > 40;
   (*inst).Solaranlage_P1 = I(int16_t,10) > I(int16_t,8);
-  if (((int64_t) ((*inst).Solaranlage_Op0)) == 0)
+  if ((*inst).Solaranlage_Op0 == ((uint8_t) (0)))
   {
     if (((*inst).Solaranlage_P0 && (*inst).Solaranlage_P1))
     {
       (*inst).Solaranlage_Op0 = ((uint8_t) (1));
     }
   }
-  else if (((int64_t) ((*inst).Solaranlage_Op0)) == 1)
+  else if ((*inst).Solaranlage_Op0 == ((uint8_t) (1)))
   {
     if ((!(*inst).Solaranlage_P0 || !(*inst).Solaranlage_P1))
     {
@@ -317,8 +317,8 @@ void Solaranlage (
   /* Initialisierung wenn keine Operation gesetzt ist */
   /* Zuweisung der Operationsvariablen */
   /* Magnetventil Fischwasser einschalten */
-  QX(0,2) = (((int64_t) ((*inst).Solaranlage_Op0)) == 1);
-  if ((((int64_t) ((*inst).Solaranlage_Op0)) == 0))
+  QX(0,2) = ((*inst).Solaranlage_Op0 == ((uint8_t) (1)));
+  if (((*inst).Solaranlage_Op0 == ((uint8_t) (0))))
   {
     (*inst).z = F(int16_t,4);
   }
@@ -356,22 +356,22 @@ void Temperatursensoren (
   /* Berechnung der nicht binären Prozeßvariablen */
   (*inst).Temperatursensoren_P0 = (*inst).diff > 10000;
   (*inst).Temperatursensoren_P1 = (*inst).diff > 10010;
-  if (((int64_t) ((*inst).Temperatursensoren_Op0)) == 0)
+  if ((*inst).Temperatursensoren_Op0 == ((uint8_t) (0)))
   {
     (*inst).Temperatursensoren_Op0 = ((uint8_t) (1));
   }
-  else if (((int64_t) ((*inst).Temperatursensoren_Op0)) == 1)
+  else if ((*inst).Temperatursensoren_Op0 == ((uint8_t) (1)))
   {
     if ((*inst).Temperatursensoren_P0)
     {
       (*inst).Temperatursensoren_Op0 = ((uint8_t) (3));
     }
   }
-  else if (((int64_t) ((*inst).Temperatursensoren_Op0)) == 2)
+  else if ((*inst).Temperatursensoren_Op0 == ((uint8_t) (2)))
   {
     (*inst).Temperatursensoren_Op0 = ((uint8_t) (0));
   }
-  else if (((int64_t) ((*inst).Temperatursensoren_Op0)) == 3)
+  else if ((*inst).Temperatursensoren_Op0 == ((uint8_t) (3)))
   {
     if ((*inst).Temperatursensoren_P1)
     {
@@ -394,12 +394,12 @@ void Temperatursensoren (
   /* Initialisierung wenn keine Operation gesetzt ist */
   /* Zuweisung der Operationsvariablen */
   /* Temperaturen vom OneWire einlesen */
-  QX(0,1) = (((int64_t) ((*inst).Temperatursensoren_Op0)) == 2);
+  QX(0,1) = ((*inst).Temperatursensoren_Op0 == ((uint8_t) (2)));
   /* Spannung für Temperatursensoren einschalten */
-  QX(0,6) = (((int64_t) ((*inst).Temperatursensoren_Op0)) == 2) || (((int64_t) ((*inst).Temperatursensoren_Op0)) == 3);
+  QX(0,6) = ((*inst).Temperatursensoren_Op0 == ((uint8_t) (2))) || ((*inst).Temperatursensoren_Op0 == ((uint8_t) (3)));
   /* Differenz berechnen */
-  _OV3 = (((int64_t) ((*inst).Temperatursensoren_Op0)) == 1) || (((int64_t) ((*inst).Temperatursensoren_Op0)) == 3);
-  if ((((int64_t) ((*inst).Temperatursensoren_Op0)) == 0))
+  _OV3 = ((*inst).Temperatursensoren_Op0 == ((uint8_t) (1))) || ((*inst).Temperatursensoren_Op0 == ((uint8_t) (3)));
+  if (((*inst).Temperatursensoren_Op0 == ((uint8_t) (0))))
   {
     F(int16_t,8) = F(int16_t,2);
   }
